+++ packages/common/src/unplugin.ts
@@ -2,25 +2,10 @@ import {
   type FilterPattern,
   createFilter as createRollupFilter,
 } from '@rollup/pluginutils'
-import { type MagicStringBase } from 'magic-string-ast'
+import { generateTransform } from 'magic-string-ast'
 
-export function getTransformResult(
-  s: MagicStringBase | undefined,
-  id: string
-): { code: string; map: any } | undefined {
-  if (s?.hasChanged()) {
-    return {
-      code: s.toString(),
-      get map() {
-        return s.generateMap({
-          source: id,
-          includeContent: true,
-          hires: true,
-        })
-      },
-    }
-  }
-}
+/** @deprecated use `generateTransform` instead */
+export const getTransformResult = generateTransform
 
 export interface BaseOptions {
   include?: FilterPattern