+++ packages/chain-call/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vue-macros/chain-call
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [[`0a1cbf79`](https://github.com/vue-macros/vue-macros/commit/0a1cbf796f00e774ba48ff019fec0e008e311faa)]:
+  - @vue-macros/common@1.6.2
+
+## 0.1.1
+
+### Patch Changes
+
+- [`07a311c7`](https://github.com/vue-macros/vue-macros/commit/07a311c7c21894cea6c7cfdff703cd044a4445a2) Thanks [@sxzz](https://github.com/sxzz)! - improve sourcemap hires
+
+- Updated dependencies [[`07a311c7`](https://github.com/vue-macros/vue-macros/commit/07a311c7c21894cea6c7cfdff703cd044a4445a2)]:
+  - @vue-macros/common@1.6.1
+
 ## 0.1.0
 
 ### Minor Changes