+++ packages/macros/CHANGELOG.md
@@ -1,5 +1,59 @@
 # unplugin-vue-macros
 
+## 2.4.2
+
+### Patch Changes
+
+- Updated dependencies [[`0a1cbf79`](https://github.com/vue-macros/vue-macros/commit/0a1cbf796f00e774ba48ff019fec0e008e311faa)]:
+  - @vue-macros/common@1.6.2
+  - @vue-macros/better-define@1.6.8
+  - @vue-macros/chain-call@0.1.2
+  - @vue-macros/define-emit@0.1.12
+  - @vue-macros/define-models@1.0.12
+  - unplugin-vue-define-options@1.3.14
+  - @vue-macros/define-prop@0.2.3
+  - @vue-macros/define-props@1.0.16
+  - @vue-macros/define-props-refs@1.1.6
+  - @vue-macros/define-render@1.3.15
+  - @vue-macros/define-slots@1.0.11
+  - @vue-macros/export-expose@0.0.9
+  - @vue-macros/export-props@0.3.14
+  - @vue-macros/hoist-static@1.4.8
+  - @vue-macros/jsx-directive@0.2.1
+  - @vue-macros/named-template@0.3.15
+  - @vue-macros/reactivity-transform@0.3.18
+  - @vue-macros/setup-block@0.2.14
+  - @vue-macros/setup-component@0.16.15
+  - @vue-macros/setup-sfc@0.15.15
+  - @vue-macros/short-emits@1.4.6
+
+## 2.4.1
+
+### Patch Changes
+
+- Updated dependencies [[`623045a4`](https://github.com/vue-macros/vue-macros/commit/623045a4c977273939f3109844112f4d95e4e835), [`07a311c7`](https://github.com/vue-macros/vue-macros/commit/07a311c7c21894cea6c7cfdff703cd044a4445a2)]:
+  - @vue-macros/jsx-directive@0.2.0
+  - @vue-macros/reactivity-transform@0.3.17
+  - @vue-macros/define-props-refs@1.1.5
+  - @vue-macros/setup-component@0.16.14
+  - unplugin-vue-define-options@1.3.13
+  - @vue-macros/named-template@0.3.14
+  - @vue-macros/better-define@1.6.7
+  - @vue-macros/define-models@1.0.11
+  - @vue-macros/define-render@1.3.14
+  - @vue-macros/export-expose@0.0.8
+  - @vue-macros/define-props@1.0.15
+  - @vue-macros/define-slots@1.0.10
+  - @vue-macros/export-props@0.3.13
+  - @vue-macros/hoist-static@1.4.7
+  - @vue-macros/define-emit@0.1.11
+  - @vue-macros/define-prop@0.2.2
+  - @vue-macros/setup-block@0.2.13
+  - @vue-macros/short-emits@1.4.5
+  - @vue-macros/chain-call@0.1.1
+  - @vue-macros/setup-sfc@0.15.14
+  - @vue-macros/common@1.6.1
+
 ## 2.4.0
 
 ### Minor Changes